requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
